Move SwcComments
